requirements.txt
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
fonttools==4.41.1
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
seaborn==0.12.2
six==1.16.0
tqdm==4.65.0
tzdata==2023.3
